+++ os/src/task/task.rs
@@ -73,7 +73,7 @@ pub struct TaskControlBlockInner {
     pub priority: u8,
 
     /// stride
-    pub stride: u16,
+    pub stride: u32,
 }
 
 impl TaskControlBlockInner {
@@ -248,7 +248,7 @@ impl TaskControlBlock {
     }
 }
 
-#[derive(Copy, Clone, PartialEq)]
+#[derive(Copy, Clone, PartialEq, Debug)]
 /// task status: UnInit, Ready, Running, Exited
 pub enum TaskStatus {
     /// uninitialized
